+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_functions.go
@@ -475,7 +475,7 @@ func (backend *DockerKurtosisBackend) GetConnectionWithUserService(
 }
 
 // It returns io.ReadCloser which is a tar stream. It's up to the caller to close the reader.
-func (backend *DockerKurtosisBackend) CopyFromUserService(
+func (backend *DockerKurtosisBackend) CopyFilesFromUserService(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
 	serviceGuid service.ServiceGUID,